feat: Disable changes to Emacs Meta keybinding
This will allow me to jump forward/backward by words when using
insert mode, or deleting words backwards (M-f) when opening files.

Will continue to do this until I get used to Xah Fly keys.

Also changed CapsLock key to be Home key via Xmodmap along with
this change, as M-Space was not flowing as well. One key for insert
mode 'u' and one key for command mode 'Home', which is in now in
the same row on my left hand.